Initial Check Satisfied TestPerson, Implemented Average Gallery

+++ strong_sort/reid_multibackend.py
@@ -135,7 +135,7 @@ class ReIDDetectMultiBackend(nn.Module):
 
     def preprocess(self, im_crops):
         def _resize(im, size):
-            return cv2.resize(im.astype(np.float32), size)
+            return cv2.resize(im, size)
 
         im = torch.cat([self.norm(_resize(im, self.size)).unsqueeze(0) for im in im_crops], dim=0).float()
         im = im.float().to(device=self.device)
